Add files via upload
Change a new way to calculate Gr and add Lr calculation. The input of calculations now using the distance from the reference point rather than the coordinate of the points. It improve the resolution and accuracy of  both Gr and Lr function. The plot code also change to fit with new output data files